% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ShrinkNet-methods.R
\docType{class}
\name{ShrinkNet-class}
\alias{ShrinkNet-class}
\alias{print,ShrinkNet-method}
\alias{print}
\alias{show,ShrinkNet-method}
\alias{show}
\alias{summary,ShrinkNet-method}
\alias{summary}
\alias{adjacency,ShrinkNet-method}
\alias{adjacency}
\alias{score,ShrinkNet-method}
\alias{score}
\alias{plotML,ShrinkNet-method}
\alias{plotML}
\alias{plotPrior,ShrinkNet-method}
\alias{plotPrior}
\alias{plotGraph,ShrinkNet-method}
\alias{plotGraph}
\alias{listEdges,ShrinkNet-method}
\alias{listEdges}
\alias{topEdges,ShrinkNet-method}
\alias{topEdges}
\alias{topDegree,ShrinkNet-method}
\alias{topDegree}
\alias{coef,ShrinkNet-method}
\alias{coef}
\title{Class ShrinkNet}
\usage{
\S4method{print}{ShrinkNet}(x, ...)

\S4method{show}{ShrinkNet}(object)

\S4method{summary}{ShrinkNet}(object, ...)

\S4method{adjacency}{ShrinkNet}(object)

\S4method{score}{ShrinkNet}(object)

\S4method{plotML}{ShrinkNet}(object, ...)

\S4method{plotPrior}{ShrinkNet}(object, ...)

\S4method{plotGraph}{ShrinkNet}(object, ...)

\S4method{listEdges}{ShrinkNet}(object)

\S4method{topEdges}{ShrinkNet}(object, nb = 20)

\S4method{topDegree}{ShrinkNet}(object, nb = 20)

\S4method{coef}{ShrinkNet}(object, ...)
}
\arguments{
\item{x}{An object of class \code{ShrinkNet-class}}

\item{...}{further arguments passed to or from other methods.}

\item{object}{An object of class \code{ShrinkNet-class}}

\item{nb}{maximum number of edges or nodes (depending on
the function it applies to) to return. Default is 20.}
}
\description{
An S4 class representing the output of
the \code{\link{ShrinkNet}} function.
}
\section{Slots}{

\describe{
\item{\code{graph}}{\code{\link[igraph]{igraph}} object}

\item{\code{kappa}}{\code{matrix} of edge ranking scores}

\item{\code{beta}}{\code{matrix} containing the posterior mean of
the SEM's regression coefficients (the ith column
corresponds to the ith regression equation).}

\item{\code{p0}}{\code{numeric}, estimate of the
proportion of null hypothesis.}

\item{\code{logMaxBFs}}{\code{matrix} containing the log-max
Bayes factor for each edge (see Leday et. al (2015)).}

\item{\code{globalPrior}}{\code{matrix} of the
(shape and rate) parameter values for the global shrinkage prior
(gamma distribution) at each iteration of the variational algorithm.}

\item{\code{allmargs}}{\code{matrix} of variational lower bounds for
each regression equation in the SEM and each iteration of
the variational algorithm.}

\item{\code{time}}{\code{numeric}, running time (in seconds).}
}}

\references{
Leday, G.G.R., de Gunst, M.C.M., Kpogbezan, G.B.,
van der Vaart, A.W., van Wieringen, W.N., and
van de Wiel, M.A. (2015). Gene network reconstruction using
global-local shrinkage priors. The Annals of Applied Statistics.
11 (2017), no. 1, 41--68.
}
\author{
Gwenael G.R. Leday <gwenael.leday (at) mrc-bsu.cam.ac.uk>
}
