% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShrinkNet.R
\name{ShrinkNet}
\alias{ShrinkNet}
\title{Gene network reconstruction using global-local shrinkage priors}
\usage{
ShrinkNet(
  tX,
  globalShrink = 1,
  nsamp0 = NULL,
  blfdr = 0.1,
  maxNbEdges = NULL,
  maxiter = 100,
  tol = 0.001,
  verbose = TRUE,
  standardize = TRUE
)
}
\arguments{
\item{tX}{\code{matrix}. Gene expression data with \eqn{p} rows
(markers) and \eqn{n} columns (obervations).}

\item{globalShrink}{\code{integer}. Type of global shrinkage,
equal to 1 or 2. See Details.}

\item{nsamp0}{\code{integer}. Number of randomly selected
edges used to estimate \eqn{p_0}. See Details.}

\item{blfdr}{\code{numeric}. Bayesian analogue of the
local false discovery rate used for
edge selection (value between 0 and 1). Default is 0.1.}

\item{maxNbEdges}{\code{numeric}. Maximum number of edges to select.}

\item{maxiter}{\code{integer}. Maximum number of iterations
for the variational algorithm. Default is 100.}

\item{tol}{\code{numeric}. Maximum relative convergence tolerance
over the \eqn{p} variational lower bounds. Default is 0.001.}

\item{verbose}{\code{logical}. Should information on progress be printed?}

\item{standardize}{\code{logical}. Should the data be
standardized? Default is TRUE.}
}
\value{
An object of class \code{\link{ShrinkNet-class}}
}
\description{
Gene network reconstruction using global-local shrinkage priors
}
\details{
When \code{globalShrink} = 1, empirical Bayes for the global
shrinkage prior is carried out using fixed-point iterations as
in Valpola and Honkela (2006). Otherwise, when \code{globalShrink} = 2,
the approximate analytical solution of Leday et al (2015) is used.

When \code{nsamp0} = \code{NULL} (default), the proportion of null
hypotheses \eqn{p_0} is estimated using the Bayes factors
calculated for all \eqn{P = 0.5 * p * (p - 1)} edges,
as in Leday et al (2015). However, when \eqn{P} is very large,
it may be preferable (for computational reasons) to
estimate \eqn{p_0} from a random subset of edges.
The argument \code{nsamp0} allows to specify the size of
the random subset. Currwently, the default is
\code{nsamp0} = \code{NULL} when \eqn{p <= 100} and
\code{nsamp0} = 1000 otherwise.
}
\references{
Leday, G.G.R., de Gunst, M.C.M., Kpogbezan, G.B.,
van der Vaart, A.W., van Wieringen, W.N., and
van de Wiel, M.A. (2015). Gene network reconstruction using
global-local shrinkage priors. The Annals of Applied Statistics.
11 (2017), no. 1, 41--68.
}
\author{
Gwenael G.R. Leday
}
