% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSVD}
\alias{getSVD}
\title{Convenience function for singular value decomposition}
\usage{
getSVD(ii, tX)
}
\arguments{
\item{ii}{integer. Gene index.}

\item{tX}{p by n matrix of gene expression.}
}
\value{
A named list with the following elements:
 \item{u}{A matrix containing the left singular vectors.}
 \item{d}{A vector containing the singular values.}
 \item{v}{A matrix containing the right singular vectors.}
}
\description{
Convenience function for singular value decomposition
}
\details{
The function returns the singular value decomposition of
\eqn{X_{ii} = UDV^T}, where \eqn{X_{ii}} is the transpose of
the matrix \code{tX} without the \code{ii}th row.
}
