% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r, R/ShrinkNet-methods.r
\docType{class}
\name{ShrinkNet-class}
\alias{ShrinkNet-class}
\alias{adjacency}
\alias{adjacency,ShrinkNet-method}
\alias{listEdges}
\alias{listEdges,ShrinkNet-method}
\alias{plotGraph}
\alias{plotGraph,ShrinkNet-method}
\alias{plotML}
\alias{plotML,ShrinkNet-method}
\alias{plotPrior}
\alias{plotPrior,ShrinkNet-method}
\alias{print}
\alias{print,ShrinkNet-method}
\alias{score}
\alias{score,ShrinkNet-method}
\alias{show}
\alias{show,ShrinkNet-method}
\alias{summary}
\alias{summary,ShrinkNet-method}
\alias{topDegree}
\alias{topDegree,ShrinkNet-method}
\alias{topEdges}
\alias{topEdges,ShrinkNet-method}
\title{Class ShrinkNet}
\usage{
\S4method{print}{ShrinkNet}(x, ...)

\S4method{show}{ShrinkNet}(object)

\S4method{summary}{ShrinkNet}(object, ...)

\S4method{adjacency}{ShrinkNet}(object)

\S4method{score}{ShrinkNet}(object)

\S4method{plotML}{ShrinkNet}(object, ...)

\S4method{plotPrior}{ShrinkNet}(object, ...)

\S4method{plotGraph}{ShrinkNet}(object, ...)

\S4method{listEdges}{ShrinkNet}(object)

\S4method{topEdges}{ShrinkNet}(object, nb = 20)

\S4method{topDegree}{ShrinkNet}(object, nb = 20)
}
\arguments{
\item{x}{An object of class \code{ShrinkNet-class}}

\item{...}{further arguments passed to or from other methods.}

\item{object}{An object of class \code{ShrinkNet-class}}

\item{nb}{maximum number of edges (or nodes, depending on the function it applies to) to return, Default is 20.}
}
\description{
An S4 class representing the output of the \code{\link{ShrinkNet}} function.
}
\section{Slots}{

\describe{
\item{\code{graph}}{igraph. A graph object of class \code{\link[igraph]{igraph}}.}

\item{\code{kappa}}{matrix. The matrix containing scores used to rank edges.}

\item{\code{p0}}{numeric. A number giving the estimate of the proportion of null hypothesis.}

\item{\code{logMaxBFs}}{matrix. The matrix containing the log-max Bayes factor for each edge (see Leday et. al (2015)).}

\item{\code{globalPrior}}{matrix. A matrix containing the (shape and rate) parameters of the global shrinkage prior
(gamma distribution) at each iteration of the variational algorithm.}

\item{\code{allmargs}}{matrix. A matrix containing the values of the variational lower bounds for each regression
equation in the Bayesian SEM at each iteration of the variational algorithm.}

\item{\code{time}}{numeric. Running time (in seconds) of ShrinkNet.}
}}
\author{
Gwenael G.R. Leday <gwenael.leday (at) mrc-bsu.cam.ac.uk>
}
\references{
Leday, G.G.R., de Gunst, M.C.M., Kpogbezan, G.B., van der Vaart, A.W., van Wieringen, W.N., and
van de Wiel, M.A. (2015). Gene network reconstruction using global-local shrinkage priors. Submitted.
}

