% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShrinkNet.R
\name{ShrinkNet}
\alias{ShrinkNet}
\title{Gene network reconstruction using global-local shrinkage priors}
\usage{
ShrinkNet(tX, globalShrink = 1, methodp0 = "exact", nsamp = 1000,
  ncpus = 1, blfdr = 0.1, maxiter = 100, tol = 0.001, verbose = TRUE)
}
\arguments{
\item{tX}{p by n data matrix of scaled gene expression measurements}

\item{globalShrink}{integer. Either 1 or 2. See Details.}

\item{methodp0}{character. Either "exact" or "sampling". See Details.}

\item{nsamp}{integer. Number of randomly selected edges to estimate p0. Only when methodp0="sampling". See Details.}

\item{ncpus}{integer. Number of cpu cores to be used.}

\item{blfdr}{numeric. Bayesian analogue of the local false discovery rate used for edge selection. Value should be between 0 and 1. Default is 0.1.}

\item{maxiter}{integer. Maximum number of iterations for the variational algorithm. Default is 100.}

\item{tol}{numeric. Represents the maximum relative convergence tolerance over the p variational lower bounds. Default is 0.001.}

\item{verbose}{logical. Should information on progress be printed?}
}
\value{
A named list with the following elements:
 \item{adjacency}{A sparse matrix of class \code{\link[Matrix]{dsCMatrix-class}} containing the adjacency matrix corresponding to the selected graph.}
 \item{p0}{A number giving the estimate of the proportion of null hypothesis.}
 \item{kappabar}{A matrix containing scores used to rank edges.}
 \item{globalPrior}{A matrix containing the (shape and rate) parameters of the global shrinkage prior
 (gamma distribution) at each iteration of the variational algorithm.}
 \item{allmargs}{A matrix containing the values of the variational lower bounds for each regression
 equation in the Bayesian SEM at each iteration of the variational algorithm.}
 \item{time}{Running time of ShrinkNet.}
}
\description{
Gene network reconstruction using global-local shrinkage priors
}
\details{
<<<<<<< HEAD
The function enables the reconstruction an undirected network given data.
=======
The function enables the reconstruction of an undirected network given data.
>>>>>>> 1e661989d45a6a528a695a52841e919e34f8c7a5
Although the tool was primarily developed to analyse mRNA expression data,
the method is general and can be applied to any data set for which it is reasonable to
assume a multivariate Gaussian model. This typically include molecular data generated from
a microarray-based technology such as protein expression data (e.g. as produced by reverse-phase protein arrays)
and microRNA data.

\code{ShrinkNet} aims to be computationally efficient. Core functions are implemented in C++ using
the Rcpp and RcppArmadillo software packages and SVD decompositions are employed to speed up the
variational algorithm. Furthermore, functions in the package has been designed so that the most
computationally intensive steps can be parallelized.

If \code{globalShrink}=1 then empirical Bayes for the global shrinkage prior is carried out using
fixed-point iterations as in Valpola and Honkela (2006). Otherwise, if \code{globalShrink}=2,
the approximate analytical solution of Leday et al (2015) is used.

If \code{methodp0}="exact", all P=0.5*p*(p-1) edges are investigated whereas if \code{methodp0}="sampling"
a random subset of size \code{nsamp} is selected to estimate p0. When p<=100 we suggest using \code{methodp0}="exact" and
\code{methodp0}="sampling" otherwise with \code{nsamp} being greater than 1000.
}
\author{
Gwenael G.R. Leday <gwenael.leday (at) mrc-bsu.cam.ac.uk>
}

