% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShrinkNet.R
\name{ShrinkNet}
\alias{ShrinkNet}
\title{Gene network reconstruction using global-local shrinkage priors}
\usage{
ShrinkNet(tX, globalShrink = 1, nsamp0 = NULL, blfdr = 0.1,
  maxiter = 100, tol = 0.001, verbose = TRUE, standardize = TRUE)
}
\arguments{
\item{tX}{p by n data matrix of gene expression measurements (centered and scaled)}

\item{globalShrink}{integer. Either 1 or 2. See Details.}

\item{nsamp0}{integer. Number of randomly selected edges to estimate p0. See Details.}

\item{blfdr}{numeric. Bayesian analogue of the local false discovery rate used for edge selection. Value should be between 0 and 1. Default is 0.1.}

\item{maxiter}{integer. Maximum number of iterations for the variational algorithm. Default is 100.}

\item{tol}{numeric. Represents the maximum relative convergence tolerance over the p variational lower bounds. Default is 0.001.}

\item{verbose}{logical. Should information on progress be printed?}

\item{standardize}{logical. Should the data be standardized? Default is TRUE.}
}
\value{
An object of class \code{\link{ShrinkNet-class}}
}
\description{
Gene network reconstruction using global-local shrinkage priors
}
\details{
If \code{globalShrink}=1 then empirical Bayes for the global shrinkage prior is carried out using
fixed-point iterations as in Valpola and Honkela (2006). Otherwise, if \code{globalShrink}=2,
the approximate analytical solution of Leday et al (2015) is used.

When \code{nsamp0}=\code{NULL}, the proportion of null hypotheses p0 is estimated using Bayes factors calculated for all P=0.5*p*(p-1) edges (cf Leday et al., 2015).
When P is very large, it may me preferable to approximate p0 instead using a random subset of edges.
If \code{nsamp0} is an integer, then a random subset of size \code{nsamp0} is selected to estimate p0.
The default is \code{nsamp0}=\code{NULL} when p<=100 and \code{nsamp0}=1000 otherwise.
}
\author{
Gwenael G.R. Leday <gwenael.leday (at) mrc-bsu.cam.ac.uk>
}
\references{
Leday, G.G.R., de Gunst, M.C.M., Kpogbezan, G.B., van der Vaart, A.W., van Wieringen, W.N., and
van de Wiel, M.A. (2015). Gene network reconstruction using global-local shrinkage priors. Submitted.
}

