% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShrinkNet.R
\name{ShrinkNet}
\alias{ShrinkNet}
\title{Undirected network inference using a Bayesian SEM and a posteriori edge selection}
\usage{
ShrinkNet(tX, globalShrink = 1, maxiter = 100, blfdr = 0.1,
  maxedges = NULL)
}
\arguments{
\item{tX}{p by n data matrix}

\item{globalShrink}{either 1 or 2. See Details.}

\item{maxiter}{integer. Maximum number of iterations for the variational algorithm. Default is 100.}

\item{blfdr}{Bayesian analogue of the local false discovery rate used for edge selection. Value should be between 0 and 1. Default is 0.1.}

\item{maxedges}{integer. Maximum number of edges to consider for forward selection. Default is 0.25*p*(p-1). See Details.}
}
\value{
list
}
\description{
Undirected network inference using a Bayesian SEM and a posteriori edge selection
}
\details{
If \code{globalShrink}=1 then empirical Bayes for the global shrinkage prior is carried out using
fixed-point iterations as in Valpola and Honkela (2006). Otherwise, if \code{globalShrink}=2,
the approximate analytical solution of Leday et al (2015) is used.

The edge selection procedure respects the ranking on edges in terms of the order in which
they are considered in the forward selection. However, the Bayes factor generally decreases
with the rank. For this reason, it may be pratical to avoid investigating all P=0.5*p*(p-1)
possible edges since a large part of them are not important. The argument \code{maxedges}
specifies the maximum number of edges to consider both for calculating p0 and the forward
selection procedure. The default is 0.5*P.
}
\author{
Gwenael G.R. Leday <gwenael.leday (at) mrc-bsu.cam.ac.uk>
}

